SVR-73 Improve BadgeType enumeration and determination
- Refactors BadgeType into sub-package.
